study/mysql-connection/backend/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.25